Fixed to compile with dmd2.063, and run in x64 mode! Since randomness works differently in 32/64, the standard map is now updated to be the 64-bit one. Should still be runnable on 32-bits since we save the original heightmap now. Fixed so that the seed values are saved directly on world creation. Moved the gameroot-detecting unittest which actually checks that util.filesystem.dir works. Now resides in util.filesystem instead of util.socket. Also prints current working directory, to make it easier to realize you dont run from the gameroot. Fixed 'bug' with dmd64 not handling exceptions properly; setting thread name in 64-bit now checks for IsDebuggerPresent before trying; makes it possible to run without debugging it :P

+++ cgy/gui/joinmenu.d
@@ -62,7 +62,7 @@ class JoinMenu : GuiElementWindow {
 
     void onJoinGame() {
         server = servEdit.getText();
-        int oldIdx = countUntil(g_serverList, server);
+        size_t oldIdx = countUntil(g_serverList, server);
         while(oldIdx != -1) {
             g_serverList = g_serverList.remove(oldIdx);
             oldIdx = countUntil(g_serverList, server);
